Removed interface IndicatorReportLineDAO
